chore: move document types to own file
