+++ packages/playground/src/api/documents/previewTypes.ts
@@ -1,4 +1,3 @@
-
 export interface DocumentPreviewTemporaryLink {
   temporaryLink: string;
 }
